Stopped using channels in threads for no reason. Better error handle.
Loops are now fully blocking and will blow up for an exception,
which is what I want. Added job failing semantics to job queue with
limited retries. Job failing is a bit weird now due to the problems
with mapcat fanning job IDs out to multiple stages of a transduce
as part of a single pipeline. This was already a known issue.